Refactor : 레포지토리 방식에서 일급 컬렉션 방식으로 변경
1. 유틸클래스에 있던 메서드들을  LottoCollection으로 옮겼습니다.
2. 유틸클래스 테스트코드들을 삭제했습니다.
3. 서비스에서도 LottoCollection가 처리할만한 메서드를 찾아 옮겼습니다. (makeLotto)
4. 서비스는 더이상 레포지토리를 주입받지 않으며 LottoCollection를 필드로 가집니다.
