+++ src/main/java/lotto/config/AppConfig.java
@@ -2,7 +2,6 @@ package lotto.config;
 
 import lotto.controller.LottoController;
 import lotto.controller.view.LottoView;
-import lotto.repository.LottoRepository;
 import lotto.service.LottoService;
 import lotto.validate.Validators;
 
@@ -10,8 +9,7 @@ public class AppConfig {
 
 
     private final Validators validator = new Validators();
-    private final LottoRepository repository = new LottoRepository();
-    private final LottoService service = new LottoService(validator,repository);
+    private final LottoService service = new LottoService(validator);
     private final LottoView view = new LottoView();
     private final LottoController controller = new LottoController(view,service);
